Add initial support for the knockout matches.
This adds a new /comp/knockout page which displays a tree-ish layout
of the knockout matches. Each round of the knockouts becomes a row
in the tree, with the finals at the bottom.

This is then styled such that the matches are visually paired in
line with which occur at the same time. The more interesting
matches nearer the bottom are consecutively larger.

Change-Id: Id2330139c09d00ea48e97b8b0dc23cd15b873b24