Make it possible to hide old matches in the schedule.
This adds a checkbox to the schedule page allowing the user to
hide old matches. We define old matches as those which started more
than 15 minutes ago. To allow the preference to persist, we store
the value of the checkbox in a cookie (and load it when the page
loads). Default this to on as most users aren't going to care about
the older matches.

Change-Id: Ie58f29365d33f467529bf4d62fc05c4fc7c04708