Improve the knockouts layout
Add a new 'match' layer of separation in each round, as distinct
from a 'game'. A match is a collection of games which occur at the
same time. A game is something which occurs only in a single arena.

This makes styling the matches simpler, as we can now address them
as a single html entity. It also makes it much easier to think about
the things being displayed.

Change-Id: I3ecc338b8688eb1490f81b1eebfd1f45dd15fb67