Reduce polling by checking a state value.
Introduce a new State resource which uses the /state endpoint to
know when the underlying competition state has changed. While this
doesn't cover changes to the 'current' match, it does include all
other relevant changes (including match timings).

Use this on most of the pages to reduce the most frequent polling
to be this check and only when that changes update the rest of the
data. Once again the LeaguePoints controller is simple enough that
it doesn't need this.

Change-Id: I553da14755f9e9bf19b8992341a216208e3fa9e6