Ensure that clearing the list of matches works.
Previously, if the list of requested matches was empty, this was
treated as an invalid input and nothing happened. This is however
wrong since selecting the no team option of the drop down results
in clearing the list. Either that or manually clearing the list that
already had items loaded resulted in the page leaving the previous
matches showing, which was clearly wrong.

Ensure that we consider empty input as valid and clear the list of
matches if the user wants that.

Change-Id: I1a20b47af0337e71cafcbcd12ac723b81da1c96e