Ensure the Comp Team page's current session is updated propertly
Previously, this was only updated when the current match changed.
While the sessions being updated is less likelly to affect this,
in the case where the current match loaded first it meant that the
page sat for about 10s until the next current match update before
the current session information was set (since sessions had been
null the first time through). Avoid that by having both refresh it.

Change-Id: I39bd44974530d97682db334185753805df655eb5