Implement Rich's front page design.
This adds a local template for the match infos and then uses that
for the current and a couple of upcoming matches. We move the twitter
stream down as that's not very interesting really.

The only tweak over Rich's design that I've done here is to put the
table header cells in the middle, rather than on the left, which I
think works well.

Also move the leaderboard slightly to the right so that the top stuff
doesn't end up looking like it's much wider than it should be.

Change-Id: I8bef259e80f36e24df2b1227523d48c6a45ead48