Separate a Team's view of their knockout score from the league
Since we now have the scores for the knockout matches too, these
will appear in the table. However the total number of game and
league points reported is (for the moment correctly) only the total
for the league. As a result, we need to separate the knockout
matches from the league ones to avoid people thinking the system
is broken since the totals don't add up.

Introduce a new Knockout points table above the League points one
and have it only appear once there are knockout matches to show.

Change-Id: I07bc6b4d686cf1e765da3cee450ce0d337ec8a3e