Tweak score table vertical alignment
Since there are now three rows in the homepage score tables, it's
somewhat more obvious that the table rows don't quite line up with
the headings. This is a result of the headings actually being in
a separate table, but is visually annoying. By introducing some
additional padding at the top of the second and later cells we
compensate for the headings having a top border which is 1px thicker
without also making the borders of the table cells thicker.

Change-Id: Ic57a025d90c51817220da7b51cf67bf36fc4f844