Mention the staging deadline
This paragraph is deliberately somewhat vague since the actual
staging deadline is both configurable and not entirely related to
the advertised match times (which are the slots, not the actual
game start times).

I'm also about to extend the match countdown to provide live data
to the teams, which is more likely to be useful in detail, though
this general introduction is needed to give it context. It will
also both match the above timing note and the configuration properly.

Change-Id: I8294ebbc0233d6f8efb0d483c70e59a775b9f9ba