Fix bug where a team's next match didn't move onwards
This was probably introduced during the work to move to the 2015
style API. The issue was that the 'next_match' value was only being
set when the team changed or when the state changed (including the
first load). As a result, when a team's match finished the page
didn't increment to the next one.

This patch moves the next_match determining to a helper function
and then uses that both when the data changes but also as part of
the interval function which is setting the time values.

Change-Id: I09eefc700e16d6ae65c2a8f63a47e71d4c4c11d3