Restore the competition-mode homepage
This reintroduces it as a separate template file in the hope that
future iterations of the ordinary homepage won't remove it as they
have (repeatedly) in the past. This code is largely the same as it
was for the SR2015 competition (d3e8d8), with some tweaks to the
layout of the bottom of the page as we no longer have the general
boxes with information in them.

Patch set 2: also update the 'Tell me more...' link

Change-Id: I7c444f8733f2f0ab70268d0cbab8842273b6b4ef