Reintroduce LICENSE
This was removed during the extraction of these files from srweb
(https://www.studentrobotics.org/cgit/srweb.git), though deserves
to be here. This is the same license as srweb had, though with
the names of the authors replaced for those who have actually
touched these files.

This license also applies to history of the files prior to this commit.