Switch to pulling the brand in as a submodule
It turns out that GitHub pages doesn't cope with doing 'npm install'
when it builds a site. This means that we can't use npm as a means
to get the brand. Instead pull it in as a submodule and remove all
references to npm now that we don't need it.