Support Ruby 3
This ports the changes from https://github.com/srobo/website/pull/473.

This isn't officially supported by Jekyll and is a bit hacky, so
I'm deliberately not documenting the support, however as recent
Ubuntu versions now ship with Ruby 3 (not 2.7) and Jekyll isn't
showing any signs of moving towards supporting Ruby 3 we're not
left with many alternatives.